# docs/openapi.yaml
openapi: 3.1.0
info:
  title: LaForge API
  version: "3.0"
  description: >
    The HTTP API for LaForge — the one service humans, the `laforge` CLI, and
    GitHub talk to. This document is for internal reference and is not hosted
    publicly.


    **Authentication.** Two interchangeable schemes on the gated endpoints:

    - `cookieAuth` — the browser session cookie minted by the GitHub OAuth
      flow (`/auth/github/*`). Used by the web UI.
    - `bearerAuth` — `Authorization: Bearer <github-token>`, a GitHub token
      (e.g. from `laforge login`). Used by the CLI.

    Most endpoints accept either. `POST /webhook/github` is authenticated by a
    GitHub HMAC signature header instead. A handful of `/auth/*` endpoints are
    public.


    **Access levels.** Repository-scoped endpoints are gated at one of
    `read < build < manage < admin`, resolved from an instance-admin list, an
    explicit per-repository grant, or the caller's live GitHub permission on the
    repo (pull→read, push→build, admin→admin). Running or scheduling agent
    commands, power actions, rebuilds, and access overrides require `manage`.
servers:
  - url: http://localhost:8080
    description: Local development (docker compose)
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: laforge_session
    bearerAuth:
      type: http
      scheme: bearer
      description: A GitHub token (user-to-server or PAT), as sent by the CLI.
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /cert-status:
    get:
      tags: [System]
      summary: Agent CA and gateway certificate expiry.
      description: Reports expiry dates (no key material). `expiring_soon` is true within 90 days.
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/CertStatus' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /webhook/github:
    post:
      tags: [System]
      summary: GitHub App webhook receiver.
      description: Authenticated by the GitHub HMAC signature header, not a session/bearer.
      security: []
      responses: { '200': { description: Processed }, '400': { $ref: '#/components/responses/BadRequest' } }
  /agent-binary/{id}:
    get:
      tags: [Agents]
      summary: Download a deployed object's per-host agent binary (one-time token).
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: token, in: query, required: true, schema: { type: string } }
      responses: { '200': { description: Binary }, '403': { $ref: '#/components/responses/Forbidden' } }
...
  /auth/github/callback:
    get:
      tags: [Auth]
      summary: OAuth callback — exchanges the code, gates sign-in, sets the session cookie.
      security: []
      parameters:
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        '302': { description: Redirect to the UI on success }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /auth/me:
    get: { tags: [Auth], summary: The signed-in account (login, avatar, instance-admin flag, timezone)., responses: { '200': { description: OK }, '401': { $ref: '#/components/responses/Unauthorized' } } }
  /auth/me/timezone:
    post: { tags: [Auth], summary: Set the account's IANA timezone preference., responses: { '200': { description: OK }, '400': { $ref: '#/components/responses/BadRequest' } } }
...
  /home/attention/dismiss:
    post:
      tags: [Repositories]
      summary: Close one needs-attention item for the current user (per-person; persists until the build goes away).
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [build_id, category], properties: { build_id: { type: string, format: uuid }, category: { type: string, enum: [build_failed, objects_failed, agents_missing, steps_failed] } } } } } }
      responses: { '204': { description: Dismissed }, '400': { description: Unknown category or bad build_id }, '404': { description: Build not visible } }
  /home/attention/undismiss:
    post:
      tags: [Repositories]
      summary: Reopen a previously-closed needs-attention item for the current user.
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [build_id, category], properties: { build_id: { type: string, format: uuid }, category: { type: string, enum: [build_failed, objects_failed, agents_missing, steps_failed] } } } } } }
      responses: { '204': { description: Reopened }, '400': { description: Unknown category or bad build_id }, '404': { description: Build not visible } }
  /repos:
    get: { tags: [Repositories], summary: List registered repositories visible to the caller., responses: { '200': { description: OK } } }
    post: { tags: [Repositories], summary: Register a repository (requires GitHub push)., responses: { '201': { description: Created }, '403': { $ref: '#/components/responses/Forbidden' } } }
  /repos/{id}:
    get:
      tags: [Repositories]
      summary: One repository.
      parameters: [ { $ref: '#/components/parameters/RepoId' } ]
      responses: { '200': { description: OK }, '404': { $ref: '#/components/responses/NotFound' } }
  /repos/{id}/builds:
    get: { tags: [Builds], summary: Builds for a repository., parameters: [ { $ref: '#/components/parameters/RepoId' } ], responses: { '200': { description: OK } } }
  /repos/{id}/branches:
    get: { tags: [Repositories], summary: Branches of the repository (via GitHub)., parameters: [ { $ref: '#/components/parameters/RepoId' } ], responses: { '200': { description: OK } } }
  /repos/{id}/environment-files:
    get: { tags: [Repositories], summary: Environment files found in a branch., parameters: [ { $ref: '#/components/parameters/RepoId' } ], responses: { '200': { description: OK } } }
  /repos/{id}/people:
    get: { tags: [Repositories], summary: People CSV sources referenced by the repository's content., parameters: [ { $ref: '#/components/parameters/RepoId' } ], responses: { '200': { description: OK } } }
  /repos/{id}/access:
    get: { tags: [Access], summary: The repository's access list (people and effective levels)., parameters: [ { $ref: '#/components/parameters/RepoId' } ], responses: { '200': { description: OK } } }
  /repos/{id}/my-access:
    get: { tags: [Access], summary: The caller's own effective level on the repository., parameters: [ { $ref: '#/components/parameters/RepoId' } ], responses: { '200': { description: OK } } }
  /repos/{id}/access/{login}:
    put:
      tags: [Access]
      summary: Grant or change a person's access level (admin).
      parameters:
        - { $ref: '#/components/parameters/RepoId' }
        - { name: login, in: path, required: true, schema: { type: string } }
      responses: { '200': { description: OK }, '403': { $ref: '#/components/responses/Forbidden' } }
    delete:
      tags: [Access]
      summary: Remove a person's explicit grant (admin).
      parameters:
        - { $ref: '#/components/parameters/RepoId' }
        - { name: login, in: path, required: true, schema: { type: string } }
      responses: { '200': { description: OK }, '403': { $ref: '#/components/responses/Forbidden' } }
  /repos/{id}/configured-builds:
    get: { tags: [Configured Builds], summary: List a repository's configured builds., parameters: [ { $ref: '#/components/parameters/RepoId' } ], responses: { '200': { description: OK } } }
    post: { tags: [Configured Builds], summary: Create a configured build (branch + env + builder) (build)., parameters: [ { $ref: '#/components/parameters/RepoId' } ], responses: { '201': { description: Created }, '403': { $ref: '#/components/responses/Forbidden' } } }
  /configured-builds/{id}:
    get: { tags: [Configured Builds], summary: One configured build., parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ], responses: { '200': { description: OK } } }
    delete: { tags: [Configured Builds], summary: Delete a configured build (build)., parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ], responses: { '200': { description: OK } } }
  /configured-builds/{id}/auto-deploy:
    post: { tags: [Configured Builds], summary: Toggle auto-deploy following (build)., parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ], responses: { '200': { description: OK } } }
  /configured-builds/{id}/lock:
    post: { tags: [Configured Builds], summary: Toggle the competition-started lock (build)., parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ], responses: { '200': { description: OK } } }
  /configured-builds/{id}/sync:
    post: { tags: [Configured Builds], summary: Sync the latest commit for this configured build (build)., parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ], responses: { '200': { description: OK } } }
  /configured-builds/{id}/builds:
    post: { tags: [Configured Builds], summary: Create (trigger) a build from the tracked commit (build)., parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ], responses: { '201': { description: Created } } }
  /content-revisions/{id}/environments:
    get: { tags: [Configured Builds], summary: Environments found in a content revision., parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ], responses: { '200': { description: OK } } }
  /builds/{id}:
    get:
      tags: [Builds]
      summary: Build detail (teams, objects, access, status).
      parameters: [ { $ref: '#/components/parameters/BuildId' } ]
      responses: { '200': { description: OK }, '404': { $ref: '#/components/responses/NotFound' } }
  /builds/{id}/deploy:
    post: { tags: [Builds], summary: Deploy a planned build (build)., parameters: [ { $ref: '#/components/parameters/BuildId' } ], responses: { '200': { description: OK }, '409': { $ref: '#/components/responses/BadRequest' } } }
  /builds/{id}/teardown:
    post: { tags: [Builds], summary: Tear down a build's infrastructure (manage)., parameters: [ { $ref: '#/components/parameters/BuildId' } ], responses: { '202': { description: Accepted } } }
  /builds/{id}/upcoming:
    get: { tags: [Builds], summary: Pending newer commit for this build, if any., parameters: [ { $ref: '#/components/parameters/BuildId' } ], responses: { '200': { description: OK } } }
  /builds/{id}/apply-upcoming:
    post: { tags: [Builds], summary: Apply the newer commit to this build in place (build)., parameters: [ { $ref: '#/components/parameters/BuildId' } ], responses: { '200': { description: OK }, '409': { $ref: '#/components/responses/BadRequest' } } }
  /builds/{id}/drift:
    get: { tags: [Builds], summary: Drift report — hoster state vs what was deployed., parameters: [ { $ref: '#/components/parameters/BuildId' } ], responses: { '200': { description: OK } } }
  /builds/{id}/events:
    get: { tags: [Builds], summary: The build's event journal., parameters: [ { $ref: '#/components/parameters/BuildId' } ], responses: { '200': { description: OK } } }
...
  /builds/{id}/findings:
    get: { tags: [Builds], summary: Findings resolved for this build., parameters: [ { $ref: '#/components/parameters/BuildId' } ], responses: { '200': { description: OK } } }
  /builds/{id}/dashboard:
    get: { tags: [Builds], summary: Dashboard aggregates for the build., parameters: [ { $ref: '#/components/parameters/BuildId' } ], responses: { '200': { description: OK } } }
  /builds/{id}/topology:
    get: { tags: [Builds], summary: The build's resolved topology (teams, networks, members)., parameters: [ { $ref: '#/components/parameters/BuildId' } ], responses: { '200': { description: OK } } }
  /builds/{id}/artifacts:
    get: { tags: [Builds], summary: Artifacts captured for the build., parameters: [ { $ref: '#/components/parameters/BuildId' } ], responses: { '200': { description: OK } } }
  /builds/{id}/artifacts/purge:
    post: { tags: [Builds], summary: Purge the build's stored artifacts (manage)., parameters: [ { $ref: '#/components/parameters/BuildId' } ], responses: { '200': { description: OK } } }
  /builds/{id}/objects:
    get: { tags: [Objects], summary: Deployed objects for the build., parameters: [ { $ref: '#/components/parameters/BuildId' } ], responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/DeployedObject' } } } } } } }
  /builds/{id}/objects/{objectId}/events:
    get: { tags: [Objects], summary: One object's event history., parameters: [ { $ref: '#/components/parameters/BuildId' }, { $ref: '#/components/parameters/ObjectId' } ], responses: { '200': { description: OK } } }
  /builds/{id}/objects/{objectId}/heartbeats:
    get: { tags: [Objects], summary: One object's agent heartbeat history (with host metrics)., parameters: [ { $ref: '#/components/parameters/BuildId' }, { $ref: '#/components/parameters/ObjectId' } ], responses: { '200': { description: OK } } }
  /builds/{id}/objects/{objectId}/steps:
    get: { tags: [Objects], summary: One object's materialized steps and their status., parameters: [ { $ref: '#/components/parameters/BuildId' }, { $ref: '#/components/parameters/ObjectId' } ], responses: { '200': { description: OK } } }
  /builds/{id}/objects/{objectId}/render:
    get: { tags: [Objects], summary: What was rendered and delivered to the object., parameters: [ { $ref: '#/components/parameters/BuildId' }, { $ref: '#/components/parameters/ObjectId' } ], responses: { '200': { description: OK } } }
  /builds/{id}/objects/{objectId}/infra:
    get: { tags: [Objects], summary: Builder placement for the object plus the root password for hand access., parameters: [ { $ref: '#/components/parameters/BuildId' }, { $ref: '#/components/parameters/ObjectId' } ], responses: { '200': { description: OK } } }
  /builds/{id}/objects/{objectId}/config:
    get:
      tags: [Objects]
      summary: One object's full authored config (ports, env, vars, depends_on, findings).
      parameters: [ { $ref: '#/components/parameters/BuildId' }, { $ref: '#/components/parameters/ObjectId' } ]
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ObjectConfig' } } } } }
  /builds/{id}/objects/{objectId}/terminal:
    get:
      tags: [Objects]
      summary: >-
        Open an interactive root/admin shell on a host/container (WebSocket upgrade).
        Needs manage access. Binary messages carry raw PTY bytes both ways; a text
        message {"type":"resize","cols":N,"rows":N} resizes. Relayed through the
        gateway to the host agent; every session is audited and a small global
        concurrency cap applies.
      parameters: [ { $ref: '#/components/parameters/BuildId' }, { $ref: '#/components/parameters/ObjectId' } ]
      responses:
        '101': { description: Switching Protocols (WebSocket) }
        '403': { description: Cross-origin refused, or caller lacks manage access }
        '409': { description: Object is not a running host/container with a live agent }
        '429': { description: Concurrent shell-session limit reached }
        '503': { description: Interactive shell is not configured on this server }
  /builds/{id}/tasks:
    post:
      tags: [Operations]
      summary: Run an ad-hoc command on the matched hosts now (manage).
      parameters: [ { $ref: '#/components/parameters/BuildId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateAdHocTaskRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AdHocResult' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /builds/{id}/power:
    post:
      tags: [Operations]
      summary: Infrastructure power action at the hoster (manage).
      parameters: [ { $ref: '#/components/parameters/BuildId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PowerRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/PowerResponse' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /builds/{id}/rebuild:
    post:
      tags: [Operations]
      summary: Tear down and recreate a host and, optionally, its dependents (manage).
      parameters: [ { $ref: '#/components/parameters/BuildId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/RebuildRequest' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RebuildResponse' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/BadRequest' }
  /builds/{id}/teams/{team}/access:
    post:
      tags: [Operations]
      summary: Open, close, extend, or reduce a team's access window (manage).
      parameters:
        - { $ref: '#/components/parameters/BuildId' }
        - { name: team, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SetTeamAccessRequest' } } }
      responses:
        '200': { description: OK }
        '202': { description: Accepted (builder call queued) }
        '400': { $ref: '#/components/responses/BadRequest' }
  /builds/{id}/scheduled-tasks:
    get: { tags: [Schedule], summary: List scheduled tasks for the build., parameters: [ { $ref: '#/components/parameters/BuildId' } ], responses: { '200': { description: OK } } }
    post:
      tags: [Schedule]
      summary: Schedule a command with a natural-language time (manage).
      parameters: [ { $ref: '#/components/parameters/BuildId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateScheduledTaskRequest' } } }
      responses:
        '201': { description: Created }
        '400': { $ref: '#/components/responses/BadRequest' }
  /builds/{id}/scheduled-tasks/preview:
    post:
      tags: [Schedule]
      summary: Validate a `when:` expression and preview its next occurrences.
      parameters: [ { $ref: '#/components/parameters/BuildId' } ]
      responses: { '200': { description: OK } }
  /builds/{id}/scheduled-tasks/{taskId}/cancel:
    post:
      tags: [Schedule]
      summary: Cancel a scheduled task (manage).
      parameters:
        - { $ref: '#/components/parameters/BuildId' }
        - { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
      responses: { '200': { description: OK } }
...
  /builder-configs/{name}:
    get:
      tags: [Builders]
      summary: One builder config (connection fields redacted for non-admins).
      parameters: [ { name: name, in: path, required: true, schema: { type: string } } ]
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/BuilderConfig' } } } }, '404': { $ref: '#/components/responses/NotFound' } }
    post: { tags: [Builders], summary: Create a builder config (admin)., parameters: [ { name: name, in: path, required: true, schema: { type: string } } ], responses: { '201': { description: Created }, '403': { $ref: '#/components/responses/Forbidden' } } }
    put: { tags: [Builders], summary: Update a builder config (admin)., parameters: [ { name: name, in: path, required: true, schema: { type: string } } ], responses: { '200': { description: OK }, '403': { $ref: '#/components/responses/Forbidden' } } }
    delete: { tags: [Builders], summary: Delete a builder config (admin)., parameters: [ { name: name, in: path, required: true, schema: { type: string } } ], responses: { '200': { description: OK }, '403': { $ref: '#/components/responses/Forbidden' } } }
  /builder-configs/{name}/available-images:
    get: { tags: [Builders], summary: Live-discover the images on the builder's hoster (admin)., parameters: [ { name: name, in: path, required: true, schema: { type: string } } ], responses: { '200': { description: OK }, '403': { $ref: '#/components/responses/Forbidden' } } }
  /builder-configs/{name}/image-builds:
    get: { tags: [Builders], summary: Base-image build history (admin)., parameters: [ { name: name, in: path, required: true, schema: { type: string } } ], responses: { '200': { description: OK } } }
    post: { tags: [Builders], summary: Trigger a base-image (re)build (admin)., parameters: [ { name: name, in: path, required: true, schema: { type: string } } ], responses: { '202': { description: Accepted } } }
...
  /builder-connections/{id}:
    get: { tags: [Builders], summary: One stored builder connection (admin)., parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ], responses: { '200': { description: OK } } }
  /image-builds/{id}/log:
    get: { tags: [Builders], summary: One base-image build's log (admin)., parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ], responses: { '200': { description: OK } } }
...
  /registry-credentials/{host}:
    delete: { tags: [Registry], summary: Delete a registry credential (admin)., parameters: [ { name: host, in: path, required: true, schema: { type: string } } ], responses: { '200': { description: OK } } }
  /registry-credentials/{host}/test:
    post: { tags: [Registry], summary: Test a registry credential (admin)., parameters: [ { name: host, in: path, required: true, schema: { type: string } } ], responses: { '200': { description: OK } } }
...